requirements.txt
python-telegram-bot[job-queue]==21.8
requests==2.32.5
